+++ src/sa/sa_get_json.php
@@ -3,7 +3,7 @@
 // HotCRP is Copyright (c) 2006-2017 Eddie Kohler and Regents of the UC
 // Distributed under an MIT-like license; see LICENSE
 
-class GetJson_SearchAction extends SearchAction {
+class GetJsonPapers_SearchAction extends SearchAction {
     private $iszip;
     private $zipdoc;
     public function __construct($iszip) {
@@ -33,9 +33,9 @@ class GetJson_SearchAction extends SearchAction {
             if ($user->can_administer($prow, true))
                 $pj[$prow->paperId] = $ps->paper_json($prow);
             else {
-                $pj[$prow->paperId] = (object) ["pid" => $prow->paperId, "error" => "You don’t have permission to administer this paper."];
+                $pj[$prow->paperId] = (object) ["pid" => $prow->paperId, "error" => "You don't have permission to administer this paper."];
                 if ($this->iszip)
-                    $this->zipdoc->warnings[] = "#$prow->paperId: You don’t have permission to administer this paper.";
+                    $this->zipdoc->warnings[] = "#$prow->paperId: You don't have permission to administer this paper.";
             }
         $pj = array_values($ssel->reorder($pj));
         if (count($pj) == 1) {
@@ -55,5 +55,57 @@ class GetJson_SearchAction extends SearchAction {
     }
 }
 
-SearchAction::register("get", "json", SiteLoader::API_GET | SiteLoader::API_PAPER, new GetJson_SearchAction(false));
-SearchAction::register("get", "jsonattach", SiteLoader::API_GET | SiteLoader::API_PAPER, new GetJson_SearchAction(true));
+class GetJsonReviews_SearchAction extends SearchAction {
+    # tailored for Review Quality Collector. Don't remove parts.
+    function allow(Contact $user) {
+        return $user->is_manager();
+    }
+    function list_actions(Contact $user, $qreq, PaperList $pl, &$actions) {
+        $actions[] = [3070, $this->subname, "Reviews", "Reviews (JSON)"];
+    }
+    function run(Contact $user, $qreq, $ssel) {
+        $conf = $user->conf;
+        $condition = "paperId in (" . implode(",", $ssel->selection()) . ")";
+        $results = [];
+        $results["hotcrp_version"] = HOTCRP_VERSION;
+        #----- get papers:
+        $result = $user->paper_result(["paperId" => $ssel->selection(), "topics" => true, "options" => true]);
+        $papers = [];
+        $ps = new PaperStatus($user->conf, $user, ["forceShow" => true, "hide_docids" => true]);
+        while (($prow = PaperInfo::fetch($result, $user)))
+            if ($user->can_administer($prow, true))
+                $papers[$prow->paperId] = $ps->paper_json($prow);
+            else
+                $papers[$prow->paperId] = (object) ["pid" => $prow->paperId, "error" => "You don’t have permission to administer this paper."];
+        $results["papers"] = array_values($ssel->reorder($papers));
+        //----- get reviewers:
+        $thefields = "contactId, firstName, lastName, email, " .
+                     "country, roles, contactTags";
+        $query = "select $thefields from contactinfo where contactId in (" .
+            "select distinct contactId from PaperReview where $condition)";
+        $contacts = $conf->qe_raw($query);
+        $results["contacts"] = $contacts->fetch_all(MYSQLI_ASSOC);
+        //----- get review form descriptor:
+        $results["reviewform"] = $conf->review_form_json();
+        //----- get reviews:
+        $thefields = "paperId, reviewId, contactId, reviewType, " .
+            "reviewSubmitted, reviewModified, reviewOrdinal, " .
+            "overAllMerit, reviewerQualification, novelty, technicalMerit, " .
+            "interestToCommunity, longevity, grammar, likelyPresentation, " .
+            "suitableForShort, paperSummary, commentsToAuthor, commentsToPC, " .
+            "commentsToAddress, weaknessOfPaper, strengthOfPaper, " .
+            "potential, fixability, textField7, textField8, " .
+            "reviewFormat";
+        $reviews = $conf->qe_raw("select $thefields from PaperReview where $condition");
+        $results["reviews"] = $reviews->fetch_all(MYSQLI_ASSOC);
+        //----- export:
+        header("Content-Type: application/json");
+        #header("Content-Disposition: attachment; filename=" . "review-data.json");
+        echo json_encode($results, JSON_PRETTY_PRINT | JSON_PARTIAL_OUTPUT_ON_ERROR | JSON_HEX_QUOT) . "\n";
+        exit;
+    }
+}
+
+SearchAction::register("get", "json", SiteLoader::API_GET | SiteLoader::API_PAPER, new GetJsonPapers_SearchAction(false));
+SearchAction::register("get", "jsonattach", SiteLoader::API_GET | SiteLoader::API_PAPER, new GetJsonPapers_SearchAction(true));
+SearchAction::register("get", "jsonreviews", SiteLoader::API_GET | SiteLoader::API_PAPER, new GetJsonReviews_SearchAction());